ai-dev-system/requirements.txt
langchain
langchain-openai
langchain-community
langgraph
docker
python-dotenv
pytest
python-gitlab
duckduckgo-search
flake8
networkx
fastapi
uvicorn[standard]
